+++ DICOM/DICOM.Shared.projitems
@@ -183,9 +183,21 @@
     <Compile Include="$(MSBuildThisFileDirectory)Media\DicomFileScanner.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Network\AsyncManualResetEvent.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Network\Client\DicomClient.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Network\Client\DicomClientAbortState.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Network\Client\DicomClientCompletedState.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Network\Client\DicomClientConnection.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Network\Client\DicomClientConnectState.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Network\Client\DicomClientDefaults.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Network\Client\DicomClientState.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Network\Client\DicomClientIdleState.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Network\Client\DicomClientLingeringState.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Network\Client\DicomClientReleaseAssociationState.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Network\Client\DicomClientRequestAssociationState.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Network\Client\DicomClientSendingRequestsState.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Network\Client\IInitialisationWithAssociationParameters.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Network\Client\DicomClientWithAssociationState.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Network\Client\DicomClientWithConnectionState.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Network\Client\IDicomClientState.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Network\Client\IInitialisationWithConnectionParameters.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Network\DicomAssociation.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Network\DicomAssociationAbortedException.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Network\DicomAssociationRejectedException.cs" />
