+++ DICOM/Network/Client/IInitialisationWithAssociationParameters.cs
@@ -1,0 +1,9 @@
+using System.Linq;
+
+namespace Dicom.Network.Client
+{
+    public interface IInitialisationWithAssociationParameters : IInitialisationWithConnectionParameters
+    {
+        DicomAssociation Association { get; set; }
+    }
+}
