+++ DICOM/Network/Client/IInitialisationWithConnectionParameters.cs
@@ -1,0 +1,7 @@
+namespace Dicom.Network.Client
+{
+    public interface IInitialisationWithConnectionParameters
+    {
+        IDicomClientConnection Connection { get; set; }
+    }
+}
